Use correct PATH to find macOS Python
When configuring the iOS/watchOS/tvOS Python, the PATH is prepended
such that the (host) macOS Python x86_64 binary can be found. But the
`$(PROJECT_DIR)/build/macOS/python/bin` is not there yet. It is created
only later in the target
`dist/Python-$(PYTHON_VER)-$1-support.b$(BUILD_NUMBER).tar.gz`.

Until then, the macOS Python is only available in the
`$(PROJECT_DIR)/$(PYTHON_DIR-macOS)/dist/` dir. Therefore, this is the
correct dir to be used in the PATH var.