Fix error No long double representation defined
When cross compiling (on macOS) to build an ARM64 target, the
npy_fpmath.h checks for certain CPU types on macOS but it does not
check for ARM at all. This leads to the error
`error No long double representation defined`. Apple dev doc says
"In other words, long double is identical to double in iOS" [1].
Therefore, nothing special has to be defined.

[1]: https://developer.apple.com/library/archive/documentation/Xcode/Conceptual/iPhoneOSABIReference/Articles/ARM64FunctionCallingConventions.html#//apple_ref/doc/uid/TP40013702-SW1