Use macOS 11 as a build host for support packages.
macOS 11 was the first release to support Apple Silicon; there are internal checks that look for macOS 11 as a proxy for Apple Silicon support. So - even though you can build Apple Silicon builds on macOS 10.15, it's safer to build on macOS 11 to make sure those checks are enabled. See https://bugs.python.org/issue42619 for details.